server state data + persistence module refactor
