+++ src/main/kotlin/io/r/raft/machine/Role.kt
@@ -1,18 +1,23 @@
 package io.r.raft.machine
 
-import io.r.raft.log.RaftLog
+import io.r.raft.persistence.RaftLog
 import io.r.raft.protocol.Index
 import io.r.raft.protocol.NodeId
 import io.r.raft.protocol.RaftMessage
 import io.r.raft.protocol.RaftRole
+import io.r.raft.protocol.Term
 import io.r.raft.transport.RaftCluster
 
 typealias RoleTransition = suspend (RaftRole) -> Role
+
 data class ServerState(
     var commitIndex: Index,
     var lastApplied: Index,
-    var leader: NodeId? = null
+    var leader: NodeId? = null,
+    var term: Term = 0L,
+    var votedFor: NodeId? = null
 )
+
 sealed class Role {
     abstract val serverState: ServerState
     abstract val timeout: Long
